+++ src/components/i18n/index.ts
@@ -1,2 +1,0 @@
-export { EuiI18n } from './i18n';
-export { EuiI18nNumber } from './i18n_number';